# docs/swagger.yaml
basePath: /api
definitions:
  global.COMPANY:
    properties:
      active:
        type: string
      address:
        type: string
      city:
        type: string
      comment:
        type: string
      company_id:
        type: integer
      company_name:
        type: string
      country:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      email:
        type: string
      logo:
        type: string
      npwp:
        type: string
      phone:
        type: string
      registration_no:
        type: integer
      state:
        type: string
      tax:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
      website_url:
        type: string
      zip_code:
        type: string
    type: object
...
  global.MODULE:
    properties:
      active:
        type: string
      comment:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      icon:
        type: string
      irl:
        type: string
      module_id:
        type: integer
      module_name:
        type: string
      module_seq:
        type: string
      parent_module:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a Api Documentation for SIPIL 2022 with Golang Backend, Gin
    Gonic Framework, GORM MySQL, and JWT Authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API DOCUMENTATION - API FOR SIPIL 2022
  version: "1.0"
paths:
...
  /company:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua company yang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua Company
      tags:
      - Company
    post:
      consumes:
      - application/json
      description: API untuk membuat company baru.
      parameters:
      - description: Company Data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/global.COMPANY'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat company baru.
      tags:
      - Company
  /company/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete company.
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete company.
      tags:
      - Company
    get:
      consumes:
      - application/json
      description: API untuk mencari company by ID yang terdapat dalam database.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Company by ID
      tags:
      - Company
    put:
      consumes:
      - application/json
      description: API untuk update company.
      parameters:
      - description: Company Data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/global.COMPANY'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update company.
      tags:
      - Company
...
  /module:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua module yang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua Modules
      tags:
      - Modules
    post:
      consumes:
      - application/json
      description: API untuk membuat module baru.
      parameters:
      - description: Modules Data
        in: body
        name: module
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat module baru.
      tags:
      - Modules
  /module/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete module.
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete module.
      tags:
      - Modules
    get:
      consumes:
      - application/json
      description: API untuk mencari module by ID yang terdapat dalam database.
      parameters:
      - description: Modules ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Modules by ID
      tags:
      - Modules
    put:
      consumes:
      - application/json
      description: API untuk update module.
      parameters:
      - description: Modules Data
        in: body
        name: module
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update module.
      tags:
      - Modules
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"